Refactor breadcrumb story
